% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jordan_Abby_Package3.R
\name{myconstr}
\alias{myconstr}
\title{PacakageProject3}
\usage{
myconstr(x, y, alpha)
}
\arguments{
\item{x}{X variable as as vector}

\item{y}{Y variable as a vector}

\item{alpha}{}
}
\value{
A constructor
}
\description{
PacakageProject3
}
\examples{
Insert Example
}
